Refactor packages
Many classes are stored in the root package.

Move classes to other packages for clarity.
